refactor: create `command` module
To add `transfer` module.
